Updating Jam source
The changes here include a simplified Windows build process:
the makefile now has 2 MSVC sections, one 32-bit the other
64-bit.

[git-p4: depot-paths = "//guest/perforce_software/jam/src/": change = 23298]

+++ execunix.c
@@ -63,7 +63,7 @@
 # define USE_MYWAIT
 # if !defined( __BORLANDC__ ) 
 # define wait my_wait
-static int my_wait( int *status );
+static long my_wait( int *status );
 # endif
 # endif
 
@@ -73,7 +73,7 @@ static void (*istat)( int );
 
 static struct
 {
-	int	pid; /* on win32, a real process handle */
+	long	pid; /* on win32, a real process handle */
 	void	(*func)( void *closure, int status );
 	void 	*closure;
 
@@ -380,7 +380,7 @@ execwait()
 
 # ifdef USE_MYWAIT
 
-static int
+static long
 my_wait( int *status )
 {
 	int i, num_active = 0;
@@ -427,7 +427,7 @@ my_wait( int *status )
 	    if ( GetExitCodeProcess(active_handles[i], &exitcode) ) {
 		CloseHandle(active_handles[i]);
 		*status = (int)((exitcode & 0xff) << 8);
-		return (int)active_handles[i];
+		return (long)active_handles[i];
 	    }
 	}
